requirements.txt
alembic==1.14.1
black==24.2.0
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
lingvoreader==0.2.15
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycairo==1.25.1
pyglossary==4.6.1
PyGObject==3.46.0
python-dotenv==1.0.1
setuptools==69.0.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
